Fix bug in usage func with incorrect return type

+++ dynamic_array/dynamic_array.h
@@ -31,7 +31,7 @@ int _size(struct DynamicArray *da);
 int _capacity(struct DynamicArray *da);
 bool _contains(struct DynamicArray *da, int item);
 void _clear(struct DynamicArray *da);
-int _usage(struct DynamicArray *da);
+double _usage(struct DynamicArray *da);
 void _discard(struct DynamicArray *da);
 void _print(struct DynamicArray *da);
